[➕Update] Add Controller to Skeleton
스켈레톤에 누락된 컨트롤러 추가

+++ Assets/02. Scripts/Data/ToolItemData.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "ToolItemData", menuName = "New Item/Weapon", order = 0)]
 public class ToolItemData : EquipItemData
 {
-    public int maxDurability;
-    public int currentDurability;
+    public float maxDurability;
+    public float currentDurability;
 
     public LayerMask targetLayers;
     public float range;
